Add blog pages and detailed blog page 🎉
