+++ src/app/posts/page.tsx
@@ -1,33 +1,29 @@
+import { BlogCard } from "@/components/BlogCard";
 import { allBlogs } from "contentlayer/generated";
-import Balancer from "react-wrap-balancer";
-import NotFound from "@/components/NotFound";
-import { Mdx } from "@/components/Mdx";
-import { formatDate } from "@/libs/utils";
+import Link from "next/link";
 
-export async function generateStaticParams() {
-  const paths = allBlogs.map((blog) => ({ slug: blog.slug }));
-
-  return paths;
-}
-
-export default async function Blog({ params }: { params: { slug: string } }) {
-  const blog = allBlogs.find((blog) => blog.slug === params.slug);
-
-  if (!blog) {
-    return <NotFound />;
-  }
+export default function Posts() {
+  const blogs = allBlogs.sort((a, b) => {
+    if (new Date(a.publishedAt) > new Date(b.publishedAt)) {
+      return -1;
+    }
+    return 1;
+  });
 
   return (
     <section>
-      <h1 className="text-2xl font-bold tracking-tighter">
-        <Balancer>{blog.title}</Balancer>
-      </h1>
-      <div className="mb-8 mt-2 flex items-center justify-between text-sm text-neutral-600 dark:text-neutral-400">
-        <p>
-          {formatDate(blog.publishedAt)} - {blog.readingTime.text}
-        </p>
-      </div>
-      <Mdx code={blog.body.code} />
+      <ul>
+        {blogs.map((blog) => (
+          <li
+            key={blog.slug}
+            className="py-1 divide-y divide-gray-200 dark:divide-gray-700"
+          >
+            <Link href={`/posts/${blog.slug}`}>
+              <BlogCard blog={blog} />
+            </Link>
+          </li>
+        ))}
+      </ul>
     </section>
   );
 }
