+++ tsconfig.json
@@ -28,10 +28,7 @@
     "**/*.ts",
     "**/*.tsx",
     ".next/types/**/*.ts",
-    ".contentlayer/generated",
-    "src/app/posts/[slug]/page.js",
-    "src/app/posts/page.js",
-    "src/app/blog/page.js"
+    ".contentlayer/generated"
   ],
   "exclude": ["node_modules"]
 }
